requirements.txt
# Honeypot Scam Detection Agent - Dependencies
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Clients
requests==2.31.0
httpx==0.26.0

# Configuration
python-dotenv==1.0.0

# Validation
validators==0.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
